Reorganize tests (split in several files)
